Fix a couple of very minor bugs in serialice:
Correct the type of the parameter 'port' in the function
pnp_set_logical_device.  This should be a u16.  I'm not aware of this
causing a problem with any of the current mainboards as there ports
are all less than 256, but this will cause problems with mainboards
that have there PNP configuration port at 0x3F0.

Correct a comment in serialice.c.  The command "wi" is a write.

Adjust the code saving and restoring eax in start.S.  The code should
save and restore eax around the enqble of FPU and SSE.  If ENABLE_SSE
is defined the value of eax is trashed.  Again, this is mainly cosmetic
as eax is actually never used.

Acked-by: Stefan Reinauer <stepan@coresystems.de>




git-svn-id: svn://svn.coreboot.org/serialice/trunk@28 8d503c2a-b2ab-11de-a960-3b3a9759a4e4

+++ chipset.c
@@ -48,7 +48,7 @@ static inline void pnp_write_register(u16 port, u8 reg, u8 value)
 	outb(value, port + 1);
 }
 
-static inline void pnp_set_logical_device(u8 port, u8 device)
+static inline void pnp_set_logical_device(u16 port, u8 device)
 {
 	pnp_write_register(port, 0x07, device);
 }
